Handle field id for parquet protobuf
When read back a parquet file with protobuf schema, try to match
the parquet schema to protobuf with respect to field id.
If the parquet file is written with a previsous parquet-protobuf (
prior to 1.9.1), field id is not persisted in parquet file metadata.
When reading these legacy file, schema converter falls back to
matching fields by name.

Change-Id: I9a15b4d1f71d1ce9b2205442fe44f9e0777700a4
JIRA: PARQUET-951